requirements.txt
loguru~=0.7.3
ruamel.yaml~=0.18.6
base58~=2.1.1
uuid~=1.30